# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VUTTR API",
    "description": "The application of a simple repository to manage tools with their respective names, links, descriptions and tags.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Localhost server"
    }
  ],
  "paths": {
    "/tools": {
      "get": {
        "summary": "Get all registered tools",
        "responses": {
          "200": {
            "description": "Get all registered tools",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tool"
                  }
                }
              }
            }
          },
          "5XX": {
            "description": "Unexpected error."
          }
        }
      },
      "post": {
        "summary": "Create new tool",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tool"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          },
          "5XX": {
            "description": "Unexpected error."
          }
        }
      }
    },
    "/tools?tag={search}": {
      "get": {
        "summary": "Get a tool by tag",
        "parameters": [
          {
            "name": "search",
            "in": "path",
            "description": "Tool tag",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "5XX": {
            "description": "Unexpected error."
          }
        }
      }
    },
    "/tools/{id}": {
      "delete": {
        "summary": "Delete a tool by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Tool ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "5XX": {
            "description": "Unexpected error."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Tool": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "title": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tags": {
            "type": "string"
          }
        }
      }
    }
  }
}